Fix wxTreeCtrl indent regression
Try to fix the problem of #25282 (hitboxes don't overlap with the expand/collapse control),
with an ABI compatible backport.

When an image list is set, check if the indent is at least FromDIP(16). If not,
increase the indent. Taking into account any previously applied indent corrections.

This could mess up things when manually changing the indent after changing image list,
but I think this is better than the expand/collapse buttons not working as expected.